requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
pyyaml==6.0.2
mlflow==2.16.0
dvc==3.54.0
dvc-s3==3.2.2